# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Prasso Invenbin API",
        "description": "API endpoints for Prasso Invenbin",
        "contact": {
            "email": "bcp@faxt.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/bom": {
            "get": {
                "tags": [
                    "BillOfMaterials"
                ],
                "summary": "Get all bills of material",
                "description": "Retrieve a list of bills of material records with eager loading of products relationship",
                "operationId": "3fa642d56982f94413fd0c566f020cd6",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "guid": {
                                                "type": "string",
                                                "example": "abc123"
                                            },
                                            "bom_name": {
                                                "type": "string",
                                                "example": "cup with logo"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "BillOfMaterials"
                ],
                "summary": "Create a new bill of material",
                "description": "Create a new bill of material",
                "operationId": "11562d1ad7ad3a543eb7f2387046b9ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BillOfMaterialsInputOutput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BillOfMaterialsInputOutput"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/bom/{guid}": {
            "get": {
                "tags": [
                    "BillOfMaterials"
                ],
                "summary": "Get a specific bill of material",
                "description": "Display the specified resource.",
                "operationId": "86f1634bb89b7cd8b3b9f29e12559ad1",
                "parameters": [
                    {
                        "name": "guid",
                        "in": "path",
                        "description": "GUID of the bill of material",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BillOfMaterialsInputOutput"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "BillOfMaterials"
                ],
                "summary": "Update a bill of material",
                "description": "Update a bill of material",
                "operationId": "6719238a5bfb8b0acc7b5cd2345b01a4",
                "parameters": [
                    {
                        "name": "guid",
                        "in": "path",
                        "description": "GUID of the bill of material",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BillOfMaterialsInputOutput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BillOfMaterialsInputOutput"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "BillOfMaterials"
                ],
                "summary": "Delete a bill of material",
                "description": "Remove the specified resource from storage.",
                "operationId": "92cd50080b5efaee3608a430789e8b12",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the bill of material",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "BillOfMaterials deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "BillOfMaterialsInputOutput": {
                "title": "ErpBom",
                "description": "Erp Bom model",
                "properties": {
                    "id": {
                        "description": "ID of the BOM",
                        "type": "integer"
                    },
                    "erp_product_id": {
                        "description": "Foreign key ID of the associated ERP product",
                        "type": "integer"
                    },
                    "bom_name": {
                        "description": "Name of the BOM",
                        "type": "string"
                    },
                    "updated_by": {
                        "description": "Foreign key ID of the user who last updated the BOM",
                        "type": "integer"
                    },
                    "updated_at": {
                        "description": "Date and time when the BOM was last updated",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "Date and time when the BOM was created",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}